pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bensbot"
version = "0.1.0"
description = "BensBot Trading System - Automated trading platform with multi-broker support"
readme = "README.md"
authors = [
    {name = "Ben Dickinson", email = "ben@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "matplotlib>=3.4.0",
    "requests>=2.25.0",
    "fastapi>=0.68.0",
    "pydantic>=1.9.0",
    "pymongo>=4.0.0",
    "redis>=4.0.0",
    "uvicorn>=0.15.0",
    "python-dotenv>=0.19.0",
    "websockets>=10.0",
    "aiohttp>=3.8.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dashboard = [
    "streamlit>=1.10.0",
    "plotly>=5.5.0",
]
backtesting = [
    "backtrader>=1.9.0",
    "yfinance>=0.1.70",
]
dev = [
    "pytest>=6.0.0",
    "black>=22.3.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.910",
]

[project.urls]
"Homepage" = "https://github.com/TheClitCommander/BensBot"
"Bug Tracker" = "https://github.com/TheClitCommander/BensBot/issues"

[project.scripts]
bensbot = "trading_bot.main:main"
bensbot-dashboard = "trading_bot.dashboard.main:main"
bensbot-backtest = "trading_bot.backtesting.main:main"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
